parse_excitation_file method
