+++ src/ansys/aedt/core/generic/file_utils.py
@@ -26,6 +26,7 @@ import codecs
 import csv
 import fnmatch
 import json
+import math
 from pathlib import Path
 from typing import Dict
 from typing import List
@@ -517,6 +518,85 @@ def read_component_file(input_file: Union[str, Path]) -> dict:
                 variables[line_list[1]] = line_list[-2]
 
     return variables
+
+
+@pyaedt_function_handler(file_name="input_file")
+def parse_excitation_file(
+    input_file: Union[str, Path],
+    is_time_domain: bool = True,
+    x_scale: float = 1.0,
+    y_scale: float = 1,
+    impedance: float = 50.0,
+    data_format: str = "Power",
+    encoding: str = "utf-8",
+    out_mag: str = "Voltage",
+    window: str = "hamming",
+) -> Union[tuple, bool]:
+    """Parse a csv file and convert data in list that can be applied to Hfss and Hfss3dLayout sources.
+
+    Parameters
+    ----------
+    input_file : str or :class:`pathlib.Path`
+        Full name of the input file.
+    is_time_domain : bool, optional
+        Either if the input data is Time based or Frequency Based. Frequency based data are Mag/Phase (deg).
+    x_scale : float, optional
+        Scaling factor for x axis.
+    y_scale : float, optional
+        Scaling factor for y axis.
+    data_format : str, optional
+        Either `"Power"`, `"Current"` or `"Voltage"`.
+    impedance : float, optional
+        Excitation impedance. Default is `50`.
+    encoding : str, optional
+        Csv file encoding.
+    out_mag : str, optional
+        Output magnitude format. It can be `"Voltage"` or `"Power"` depending on Hfss solution.
+    window : str, optional
+        Fft window. Options are ``"hamming"``, ``"hanning"``, ``"blackman"``, ``"bartlett"`` or ``None``.
+
+    Returns
+    -------
+    tuple or bool
+        Frequency, magnitude and phase.
+    """
+    try:
+        import numpy as np
+    except ImportError:
+        pyaedt_logger.error("NumPy is not available. Install it.")
+        return False
+    input_file = Path(input_file)
+    df = read_csv_pandas(input_file, encoding=encoding)
+    if is_time_domain:
+        time = df[df.keys()[0]].values * x_scale
+        val = df[df.keys()[1]].values * y_scale
+        freq, fval = compute_fft(time, val, window)
+
+        if data_format.lower() == "current":
+            if out_mag == "Voltage":
+                fval = fval * impedance
+            else:
+                fval = fval * fval * impedance
+        elif data_format.lower() == "voltage":
+            if out_mag == "Power":
+                fval = fval * fval / impedance
+        else:
+            if out_mag == "Voltage":
+                fval = np.sqrt(fval * impedance)
+        mag = list(np.abs(fval))
+        phase = [math.atan2(j, i) * 180 / math.pi for i, j in zip(list(fval.real), list(fval.imag))]
+
+    else:
+        freq = list(df[df.keys()[0]].values * x_scale)
+        if data_format.lower() == "current":
+            mag = df[df.keys()[1]].values * df[df.keys()[1]].values * impedance * y_scale * y_scale
+        elif data_format.lower() == "voltage":
+            mag = df[df.keys()[1]].values * df[df.keys()[1]].values / impedance * y_scale * y_scale
+        else:
+            mag = df[df.keys()[1]].values * y_scale
+        mag = list(mag)
+        phase = list(df[df.keys()[2]].values)
+    return freq, mag, phase
 
 
 # CAD parsing
